Moved newGroup from Contract.module to login.module
