+++ src/app/contract/contract.routing.ts
@@ -4,7 +4,6 @@ import {NewContractComponent} from './new-contract/new-contract.component';
 import {EditContractComponent} from './edit-contract/edit-contract.component';
 import {AuthGuard} from '../shared/guards/guard.service';
 import {ExecutiveGuard} from '../shared/guards/supervisor.guard';
-import {NewGroupComponent} from './new-group/new-group.component';
 
 export const contractRoutes = RouterModule.forRoot(
   [
@@ -21,9 +20,5 @@ export const contractRoutes = RouterModule.forRoot(
       // '/:contractId' is holding our hashed value. We decide the name here.
       path: 'contracts/editContract/:contractId',
       component: EditContractComponent
-    },
-    {
-      path: 'newGroup',
-      component: NewGroupComponent
     }
   ]);
