+++ src/app/login/new-group/new-group.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import {GroupService} from '../shared/group.service';
+// import {GroupService} from '../shared/group.service';
 import {Router} from '@angular/router';
-import {UserModel} from '../shared/user.model';
+import {UserModel} from '../../contract/shared/user.model';
 
 @Component({
   selector: 'app-new-group',
@@ -15,14 +15,14 @@ export class NewGroupComponent implements OnInit {
 
   onSubmit() {this.submitted = true; }
 
-  constructor(private router: Router, private groupService: GroupService) {
+  constructor(private router: Router) {
   }
 
   ngOnInit() {
   }
 
   create(email: string) {
-    this.groupService.create(email).subscribe();
+    // this.groupService.create(email).subscribe();
     this.router.navigateByUrl('login');
   }
